docs(Modal): add jsdoc for all the props related to Modal
