+++ packages/core/src/components/ModalNew/Modal/Modal.types.tsx
@@ -9,15 +9,48 @@ export type ModalCloseEvent =
   | React.KeyboardEvent<HTMLBodyElement>;
 
 export interface ModalProps extends VibeComponentProps {
+  /**
+   * Unique identifier for the modal.
+   */
   id: string;
+  /**
+   * Controls the visibility of the modal.
+   */
   show: boolean;
+  /**
+   * Determines the width and max-height of the modal.
+   */
   size?: ModalSize;
-  closeButtonTheme?: ModalTopActionsProps["color"];
+  /**
+   * Theme color for the close button.
+   */
+  closeButtonTheme?: ModalTopActionsProps["theme"];
+  /**
+   * Accessibility label for the close button.
+   */
   closeButtonAriaLabel?: ModalTopActionsProps["closeButtonAriaLabel"];
+  /**
+   * Callback fired when the modal should close.
+   */
   onClose?: (event: ModalCloseEvent) => void;
+  /**
+   * Additional action to render in the header area.
+   */
   renderHeaderAction?: ModalTopActionsProps["renderAction"];
+  /**
+   * Reference to an element that triggered the modal, used for animations.
+   */
   anchorElementRef?: React.RefObject<HTMLElement>;
+  /**
+   * When true, prevents closing the modal when clicking the overlay ("click-outside") or pressing ESC.
+   */
   alertModal?: boolean;
+  /**
+   * Modal content.
+   */
   children: React.ReactNode;
+  /**
+   * Additional inline styles for the modal.
+   */
   style?: React.CSSProperties;
 }
