+++ packages/core/src/components/ModalNew/ModalContent/ModalContent.types.ts
@@ -2,5 +2,8 @@ import React from "react";
 import { VibeComponentProps } from "../../../types";
 
 export interface ModalContentProps extends VibeComponentProps {
+  /**
+   * Main content of the modal.
+   */
   children?: React.ReactNode;
 }
