+++ packages/core/src/components/ModalNew/ModalHeader/ModalHeader.types.ts
@@ -7,7 +7,14 @@ interface WithoutDescription {
 }
 
 interface WithDescription {
+  /**
+   * Descriptive text or content below the title.
+   * When supplied, would also add an aria-describedby attribute to the modal dialog element.
+   */
   description: string | React.ReactNode;
+  /**
+   * Icon to display before the description. Can only be passed when description is supplied.
+   */
   descriptionIcon?:
     | SubIcon
     | {
@@ -16,4 +23,11 @@ interface WithDescription {
       };
 }
 
-export type ModalHeaderProps = { title: string } & (WithoutDescription | WithDescription) & VibeComponentProps;
+export type ModalHeaderProps = {
+  /**
+   * Main heading text of the modal.
+   * When supplied, would also add an aria-labelledby attribute to the modal dialog element.
+   */
+  title: string;
+} & (WithDescription | WithoutDescription) &
+  VibeComponentProps;
