+++ packages/core/src/components/ModalNew/ModalTopActions/ModalTopActions.types.ts
@@ -3,7 +3,7 @@ import MenuButton from "../../MenuButton/MenuButton";
 import IconButton from "../../IconButton/IconButton";
 import { ButtonColor } from "../../Button/ButtonConstants";
 
-export type ModalTopActionsColor = "light" | "dark";
+export type ModalTopActionsTheme = "light" | "dark";
 export type ModalTopActionsButtonColor =
   | ButtonColor.PRIMARY
   | ButtonColor.ON_PRIMARY_COLOR
@@ -11,13 +11,22 @@ export type ModalTopActionsButtonColor =
 
 export interface ModalTopActionsProps {
   /**
-   * action can be passed either as a function or direct
-   * it allows passing back to consumer the color he chose, so he won't have to define it twice
+   * Action element or render function for the top-right area.
+   * When provided as a function, receives the current button color theme
    */
   renderAction?:
     | React.ReactElement<typeof MenuButton | typeof IconButton>
     | ((color?: ModalTopActionsButtonColor) => React.ReactElement<typeof MenuButton | typeof IconButton>);
-  color?: ModalTopActionsColor;
+  /**
+   * Color theme for the top actions
+   */
+  theme?: ModalTopActionsTheme;
+  /**
+   * Accessibility label for the close button
+   */
   closeButtonAriaLabel?: string;
+  /**
+   * Callback fired when the close button is clicked
+   */
   onClose?: (event: React.MouseEvent<HTMLButtonElement>) => void;
 }
