+++ packages/core/src/components/ModalNew/footers/ModalFooter/ModalFooter.types.ts
@@ -2,5 +2,8 @@ import React from "react";
 import { ModalFooterBaseProps } from "../ModalFooterBase/ModalFooterBase.types";
 
 export interface ModalFooterProps extends Omit<ModalFooterBaseProps, "renderAction"> {
+  /**
+   * Optional content to render on the left side of the footer.
+   */
   renderSideAction?: React.ReactNode;
 }
