+++ packages/core/src/components/ModalNew/layouts/ModalBasicLayout/ModalBasicLayout.types.ts
@@ -2,5 +2,10 @@ import React from "react";
 import { VibeComponentProps } from "../../../../types";
 
 export interface ModalBasicLayoutProps extends VibeComponentProps {
+  /**
+   * Layout children in the following order:
+   * 1. Header content
+   * 2. Main content
+   */
   children: React.ReactNode;
 }
