+++ packages/core/src/components/ModalNew/layouts/ModalFooterShadow.types.ts
@@ -1,3 +1,6 @@
 export interface ModalFooterShadowProps {
+  /**
+   * Controls the visibility of the shadow.
+   */
   show: boolean;
 }
