+++ packages/core/src/components/ModalNew/layouts/ModalMedia.types.ts
@@ -2,5 +2,8 @@ import React from "react";
 import { VibeComponentProps } from "../../../types";
 
 export interface ModalMediaProps extends VibeComponentProps {
+  /**
+   * Media content to be displayed in the modal (image, video, Lottie, etc.).
+   */
   children: React.ReactNode;
 }
